+++ Compilers/APPComp/Headers/messages.h
@@ -35,9 +35,6 @@ void parse_lang_flag(int *argc, char **argv);
     M("Erro: não rolou de abrir o arquivo '%s'!!\n", \
       "Error: couldn't open the file '%s' for the life of me!!\n")
 
-#define MSG_ERR_USELESS_PROC \
-    M("Erro: esse processador não serve pra nada. Você não tem nada útil pra fazer não?\n", \
-      "Error: this processor is totally useless. Don't you have anything actually fun to do?\n")
 
 // info messages
 #define MSG_INFO_ITR_HANDLING \